# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing leaderboards, entries, participants, and metrics",
        "title": "Leaderboard Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/leaderboard-entries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all entries/rankings for a specific leaderboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard-entries"
                ],
                "summary": "List all entries for a leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by participant ID",
                        "name": "participant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of leaderboard entries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.LeaderboardEntryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new entry/ranking in a leaderboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard-entries"
                ],
                "summary": "Create a new leaderboard entry",
                "parameters": [
                    {
                        "description": "Leaderboard entry data",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateLeaderboardEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created leaderboard entry",
                        "schema": {
                            "$ref": "#/definitions/handlers.LeaderboardEntryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Leaderboard or participant not found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/leaderboard-entries/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a leaderboard entry by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard-entries"
                ],
                "summary": "Get a leaderboard entry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Leaderboard Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leaderboard entry details",
                        "schema": {
                            "$ref": "#/definitions/handlers.LeaderboardEntryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing leaderboard entry with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard-entries"
                ],
                "summary": "Update a leaderboard entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Leaderboard Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated leaderboard entry data",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateLeaderboardEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated leaderboard entry",
                        "schema": {
                            "$ref": "#/definitions/handlers.LeaderboardEntryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a leaderboard entry by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard-entries"
                ],
                "summary": "Delete a leaderboard entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Leaderboard Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/leaderboards/{leaderboard_id}/entries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all entries/rankings for a specific leaderboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard-entries"
                ],
                "summary": "List all entries for a leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by leaderboard ID",
                        "name": "leaderboard_id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Filter by participant ID",
                        "name": "participant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of leaderboard entries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.LeaderboardEntryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new entry/ranking in a leaderboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard-entries"
                ],
                "summary": "Create a new leaderboard entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Leaderboard ID",
                        "name": "leaderboard_id",
                        "in": "path"
                    },
                    {
                        "description": "Leaderboard entry data",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateLeaderboardEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created leaderboard entry",
                        "schema": {
                            "$ref": "#/definitions/handlers.LeaderboardEntryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Leaderboard or participant not found",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CreateLeaderboardEntryRequest": {
            "type": "object",
            "required": [
                "leaderboard_id",
                "participant_id",
                "rank",
                "score"
            ],
            "properties": {
                "last_updated": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "leaderboard_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "participant_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "rank": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "score": {
                    "type": "number",
                    "example": 100.5
                }
            }
        },
...
        "handlers.LeaderboardEntryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                },
                "last_updated": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "leaderboard_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "participant_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "rank": {
                    "type": "integer",
                    "example": 1
                },
                "score": {
                    "type": "number",
                    "example": 100.5
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
...
        "handlers.UpdateLeaderboardEntryRequest": {
            "type": "object",
            "properties": {
                "last_updated": {
                    "type": "string",
                    "example": "2023-01-02T00:00:00Z"
                },
                "rank": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2
                },
                "score": {
                    "type": "number",
                    "example": 200.75
                }
            }
        },
...
        "middleware.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}